X86: Fix movabsq immediates >= 2^63 printed as decimal in ATT syntax
Fix issue #2817: In AT&T syntax, movabsq instructions with immediate
values >= 2^63 (0x8000000000000000) were incorrectly printed in decimal
instead of hexadecimal.

Root cause: The comparison 'imm > HEX_THRESHOLD' used a signed int64_t,
so values with bit 63 set appeared negative and failed the comparison.

Fix: Cast imm to uint64_t for the comparison to handle the full 64-bit
unsigned range correctly.

Note: Regression tests for exact asm_text format cannot be added because
cstest's compare_asm_text() normalizes hex/decimal representations.

+++ arch/X86/X86ATTInstPrinter.c
@@ -732,7 +732,8 @@ static void printOperand(MCInst *MI, unsigned OpNo, SStream *O)
 		case X86_INS_MOVABS:
 		case X86_INS_MOV:
 			// do not print number in negative form
-			if (imm > HEX_THRESHOLD)
+			// Use unsigned comparison to handle values >= 2^63 correctly
+			if ((uint64_t)imm > HEX_THRESHOLD)
 				SStream_concat(O, "$0x%" PRIx64, imm);
 			else
 				SStream_concat(O, "$%" PRIu64, imm);
